Add support for TransposeFoldingOutputPass.
