+++ cinn/frontend/pass/use_program_pass.h
@@ -20,3 +20,4 @@ CINN_USE_REGISTER(Decomposer)
 CINN_USE_REGISTER(RemoveIdentity)
 CINN_USE_REGISTER(TransposeFoldingInput)
 CINN_USE_REGISTER(GemmRewriter)
+CINN_USE_REGISTER(TransposeFoldingOutput)
